refactored interfaceAdapters relating to mealPlanCreation to the right folder
