+++ src/InterfaceAdapters/ViewManagerModel.java
@@ -1,4 +1,4 @@
-package InterfaceAdapter;
+package InterfaceAdapters;
 
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
